+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Prefix the change with one of these keywords:
 - _Security_: in case of vulnerabilities.
 
 ## UNRELEASED
+- Added: bullet-list style to List component 
 
 ## [0.13.2] - 2019-09-xx